Add base_page.dart.
